Get origin destination data from the 2011 Census

This function downloads a .csv file representing movement between MSOA zones in England and Wales. By default it returns national data, but region can be set to subset the output to a specific local authority or region.

get_od(
  region = NULL,
  n = NULL,
  type = "within",
  omit_intrazonal = FALSE,
  base_url = paste0("https://s3-eu-west-1.amazonaws.com/",
    "statistics.digitalresources.jisc.ac.uk", "/dkan/files/FLOW/"),
  filename = "wu03ew_v2"
)

Arguments

region
The PCT region or local authority to download data from (e.g. west-yorkshire or Leeds). See View(pct_regions_lookup) for a full list of possible region names.

n
top n number of destinations with most trips in the 2011 census within the region.

type
the type of subsetting: one of from, to or within, specifying how the od dataset should be subset in relation to the region.

omit_intrazonal
should intrazonal OD pairs be omited from result? FALSE by default.

base_url
the base url where the OD dataset is stored

filename
the name of the file to download, if not the default MSOA level data.

Details

OD datasets available include wu03uk_v3 and others listed on the Wicid website.

Examples

# \donttest{
get_od("wight", n = 3)

#> Warning: cannot open URL 'https://s3-eu-west-1.amazonaws.com/statistics.digitalresources.jisc.ac.uk/dkan/files/FLOW/wu03ew_v2/wu03ew_v2.zip': HTTP status was '403 Forbidden'

#> Error in utils::download.file(file_url, zip_file): cannot open URL 'https://s3-eu-west-1.amazonaws.com/statistics.digitalresources.jisc.ac.uk/dkan/files/FLOW/wu03ew_v2/wu03ew_v2.zip'

get_od()

#> No region provided. Returning national OD data.

#> Warning: cannot open URL 'https://s3-eu-west-1.amazonaws.com/statistics.digitalresources.jisc.ac.uk/dkan/files/FLOW/wu03ew_v2/wu03ew_v2.zip': HTTP status was '403 Forbidden'

#> Error in utils::download.file(file_url, zip_file): cannot open URL 'https://s3-eu-west-1.amazonaws.com/statistics.digitalresources.jisc.ac.uk/dkan/files/FLOW/wu03ew_v2/wu03ew_v2.zip'

get_od(filename = "wu03uk_v3")

#> No region provided. Returning national OD data.

#> Warning: cannot open URL 'https://s3-eu-west-1.amazonaws.com/statistics.digitalresources.jisc.ac.uk/dkan/files/FLOW/wu03uk_v3/wu03uk_v3.zip': HTTP status was '403 Forbidden'

#> Error in utils::download.file(file_url, zip_file): cannot open URL 'https://s3-eu-west-1.amazonaws.com/statistics.digitalresources.jisc.ac.uk/dkan/files/FLOW/wu03uk_v3/wu03uk_v3.zip'

# }