Get origin destination data from the 2011 Census

This function downloads a .csv file representing movement between MSOA zones in England and Wales. By default it returns national data, but region can be set to subset the output to a specific local authority or region.

get_od(region = NULL, n = NULL, type = "within", omit_intrazonal = FALSE)

Arguments

region
The PCT region or local authority to download data from (e.g. west-yorkshire or Leeds). See View(pct_regions_lookup) for a full list of possible region names.

n
top n number of destinations with most trips in the 2011 census within the region.

type
the type of subsetting: one of from, to or within, specifying how the od dataset should be subset in relation to the region.

omit_intrazonal
should intrazonal OD pairs be omited from result? FALSE by default.

Examples

# \donttest{
get_od("wight", n = 3)

#> Parsed with column specification:
#> cols(
#>   `Area of residence` = col_character(),
#>   `Area of workplace` = col_character(),
#>   `All categories: Method of travel to work` = col_double(),
#>   `Work mainly at or from home` = col_double(),
#>   `Underground, metro, light rail, tram` = col_double(),
#>   Train = col_double(),
#>   `Bus, minibus or coach` = col_double(),
#>   Taxi = col_double(),
#>   `Motorcycle, scooter or moped` = col_double(),
#>   `Driving a car or van` = col_double(),
#>   `Passenger in a car or van` = col_double(),
#>   Bicycle = col_double(),
#>   `On foot` = col_double(),
#>   `Other method of travel to work` = col_double()
#> )

#> Parsed with column specification:
#> cols(
#>   MSOA11CD = col_character(),
#>   MSOA11NM = col_character(),
#>   BNGEAST = col_double(),
#>   BNGNORTH = col_double(),
#>   LONGITUDE = col_double(),
#>   LATITUDE = col_double()
#> )

#> # A tibble: 3 x 18
#>   geo_code1 geo_code2   all from_home light_rail train   bus  taxi motorbike
#>   <chr>     <chr>     <dbl>     <dbl>      <dbl> <dbl> <dbl> <dbl>     <dbl>
#> 1 E02003592 E02003592  1164         0          0     1    41     3        10
#> 2 E02003589 E02003589   875         0          0     0    12     1         8
#> 3 E02003591 E02003589   852         0          0     1     6     0        11
#> # … with 9 more variables: car_driver <dbl>, car_passenger <dbl>,
#> #   bicycle <dbl>, foot <dbl>, other <dbl>, geo_name1 <chr>, geo_name2 <chr>,
#> #   la_1 <chr>, la_2 <chr>

# }